Intento de mostrar una alerta y que no se dupliquen los productos

+++ js/product-info.js
@@ -256,15 +256,30 @@ document.addEventListener("DOMContentLoaded", () => {
 });
 
  //Evento del boton "agregar al carrito" que esta junto a los productos
+const infoProducto = JSON.parse(localStorage.getItem('infoProducto')) || [];
+
 function carrito(nombre, precio, imagen){
-  const DatosProducto={
-    nombre: info.name, 
-    precio: `${info.currency} ${info.cost}`, 
-    imagen: info.images
-  };
-  console.log('Producto agregado al carrito:', DatosProducto); //Para probar si los datos se envian de forma correcta
-  const infoProducto=JSON.parse(localStorage.getItem('infoProducto')) || []; //Recuperar los datos almacenado de los productos y los convierte en objetos JavaScript. Si no hay datos, queda vacio, para evitar errores
-  infoProducto.push(DatosProducto);
-  localStorage.setItem('infoProducto', JSON.stringify(infoProducto));
+  const ProductoExistente = infoProducto.find(item => item.nombre === nombre);
+  
+  if (ProductoExistente){
+    alert ("El producto ya esta en el carrito");
+  } else{
+    const DatosProducto={
+      nombre: info.name, 
+      precio: `${info.currency} ${info.cost}`, 
+      imagen: info.images,
+      cantidad: 1
+    };
+
+    infoProducto.push(DatosProducto);
+    localStorage.setItem('infoProducto', JSON.stringify(infoProducto));
+    console.log('Producto agregado al carrito:', DatosProducto);
+  }
 }
+
+function agregarAlCarrito(nombre, precio, imagen) {
+  carrito(nombre, precio, imagen);
+}
+
+
